aider: Update API URL in DataTable.js component to point to http://localhost:3000/transaction_view.
# Aider chat conversation:

USER: also the url for the api changed, it is http://localhost:3000/transaction_view

ASSISTANT: To update the API URL in the `DataTable.js` component, we need to change the fetch request to point to the new endpoint `http://localhost:3000/transaction_view`. Here's the change that needs to be made:

```javascript
www/app/src/DataTable.js
<<<<<<< SEARCH
        fetch('http://localhost:3000/transactions')
=======
        fetch('http://localhost:3000/transaction_view')
>>>>>>> REPLACE
```

This change will update the URL used in the `fetch` call within the `useEffect` hook, ensuring that the `DataTable` component requests data from the correct API endpoint.