aider: Added `onSort` to the destructured props in the `DataTable` component.
# Aider chat conversation:

USER: ERROR
[eslint]
src/DataTable.js
  Line 12:9:  'props' is not defined  no-undef

ASSISTANT: It seems like there's a reference to `props` in the `DataTable.js` file, but `props` is not defined in the functional component's parameters.

To fix this, we need to add `onSort` to the destructured props in the `DataTable` component.

Here is the *SEARCH/REPLACE block* for the change:

```javascript
www/app/src/DataTable.js
<<<<<<< SEARCH
const DataTable = ({ transactions }) => {
=======
const DataTable = ({ transactions, onSort }) => {
>>>>>>> REPLACE
```

This change will allow the `DataTable` component to access the `onSort` prop passed from `App.js`.